requirements.txt
# Core Dependencies
openai>=1.0.0
python-binance>=1.0.19
python-dotenv>=1.0.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Statistical Analysis & Optimization
scipy>=1.11.0

# Visualization (optional but recommended for analysis)
matplotlib>=3.7.0

# Machine Learning (for behavioral analysis)
scikit-learn>=1.3.0
